api/requirements.txt
Flask
pymongo
Flask-PyMongo
python-dotenv
openai
Flask-Cors
cors
gunicorn
pyjwt
wraps